+++ modules/generation/src/main/kotlin/app/epistola/generation/pdf/ImageNodeRenderer.kt
@@ -86,6 +86,7 @@ class ImageNodeRenderer : NodeRenderer {
             context.fontCache,
             context.renderingDefaults.componentDefaults("image"),
             context.renderingDefaults.baseFontSizePt,
+            context.spacingUnit,
         )
         div.add(image)